refactor(optimizer): operator
