+++ src/query/sql/src/planner/optimizer/optimizer.rs
@@ -32,12 +32,12 @@ use crate::optimizer::cascades::CascadesOptimizer;
 use crate::optimizer::decorrelate::decorrelate_subquery;
 use crate::optimizer::distributed::optimize_distributed_query;
 use crate::optimizer::distributed::SortAndLimitPushDownOptimizer;
-use crate::optimizer::filter::DeduplicateJoinConditionOptimizer;
-use crate::optimizer::filter::PullUpFilterOptimizer;
 use crate::optimizer::hyper_dp::DPhpy;
 use crate::optimizer::ir::Memo;
 use crate::optimizer::ir::SExpr;
 use crate::optimizer::join::SingleToInnerOptimizer;
+use crate::optimizer::operator::DeduplicateJoinConditionOptimizer;
+use crate::optimizer::operator::PullUpFilterOptimizer;
 use crate::optimizer::rule::TransformResult;
 use crate::optimizer::statistics::CollectStatisticsOptimizer;
 use crate::optimizer::util::contains_local_table_scan;
